// face-check-ui/src/app/services/models/payment-history-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface PaymentHistoryResponse {
  amount?: number;
  companyId?: number;
  createDate?: string;
  notes?: string;
  paymentDate?: string;
  paymentHistoryIrsId?: number;
  paymentType?: 'PAYROLL_TAX_941' | 'UNEMPLOYMENT_TAX_940' | 'STATE_UNEMPLOYMENT_TAX' | 'PERSONAL_INSURANCE' | 'WC_Payment' | 'MCTMT_PREPAYMENT' | 'MCTMT_CREDIT';
  quarter?: number;
  totalPayedAmountForQuarter?: number;
  year?: number;
}