// face-check-ui/src/app/services/models/page-response-payment-history-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { PaymentHistoryResponse } from '../models/payment-history-response';
export interface PageResponsePaymentHistoryResponse {
  content?: Array<PaymentHistoryResponse>;
  first?: boolean;
  last?: boolean;
  number?: number;
  size?: number;
  totalElement?: number;
  totalPages?: number;
}